[REF] mail,im_livechat: move leaveChannel to channel
